+++ server/src/main/java/org/opensearch/common/network/NetworkModule.java
@@ -55,6 +55,8 @@ import org.opensearch.core.xcontent.XContentParser;
 import org.opensearch.http.HttpServerTransport;
 import org.opensearch.index.shard.PrimaryReplicaSyncer.ResyncTask;
 import org.opensearch.plugins.NetworkPlugin;
+import org.opensearch.plugins.SecureHttpTransportSettingsProvider;
+import org.opensearch.plugins.SecureSettingsFactory;
 import org.opensearch.plugins.SecureTransportSettingsProvider;
 import org.opensearch.ratelimitting.admissioncontrol.enums.AdmissionControlActionType;
 import org.opensearch.tasks.RawTaskStatus;
@@ -74,7 +76,9 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.function.Supplier;
+import java.util.stream.Collectors;
 
 /**
  * A module to handle registering and binding all network related classes.
@@ -173,13 +177,31 @@ public final class NetworkModule {
         ClusterSettings clusterSettings,
         Tracer tracer,
         List<TransportInterceptor> transportInterceptors,
-        Collection<SecureTransportSettingsProvider> secureTransportSettingsProvider
+        Collection<SecureSettingsFactory> secureSettingsFactories
     ) {
         this.settings = settings;
 
-        if (secureTransportSettingsProvider.size() > 1) {
+        final Collection<SecureTransportSettingsProvider> secureTransportSettingsProviders = secureSettingsFactories.stream()
+            .map(p -> p.getSecureTransportSettingsProvider(settings))
+            .filter(Optional::isPresent)
+            .map(Optional::get)
+            .collect(Collectors.toList());
+
+        if (secureTransportSettingsProviders.size() > 1) {
             throw new IllegalArgumentException(
-                "there is more than one secure transport settings provider: " + secureTransportSettingsProvider
+                "there is more than one secure transport settings provider: " + secureTransportSettingsProviders
+            );
+        }
+
+        final Collection<SecureHttpTransportSettingsProvider> secureHttpTransportSettingsProviders = secureSettingsFactories.stream()
+            .map(p -> p.getSecureHttpTransportSettingsProvider(settings))
+            .filter(Optional::isPresent)
+            .map(Optional::get)
+            .collect(Collectors.toList());
+
+        if (secureHttpTransportSettingsProviders.size() > 1) {
+            throw new IllegalArgumentException(
+                "there is more than one secure HTTP transport settings provider: " + secureHttpTransportSettingsProviders
             );
         }
 
@@ -213,9 +235,9 @@ public final class NetworkModule {
                 registerTransport(entry.getKey(), entry.getValue());
             }
 
-            // Register any secure transports if available
-            if (secureTransportSettingsProvider.isEmpty() == false) {
-                final SecureTransportSettingsProvider secureSettingProvider = secureTransportSettingsProvider.iterator().next();
+            // Register any HTTP secure transports if available
+            if (secureHttpTransportSettingsProviders.isEmpty() == false) {
+                final SecureHttpTransportSettingsProvider secureSettingProvider = secureHttpTransportSettingsProviders.iterator().next();
 
                 final Map<String, Supplier<HttpServerTransport>> secureHttpTransportFactory = plugin.getSecureHttpTransports(
                     settings,
@@ -233,6 +255,11 @@ public final class NetworkModule {
                 for (Map.Entry<String, Supplier<HttpServerTransport>> entry : secureHttpTransportFactory.entrySet()) {
                     registerHttpTransport(entry.getKey(), entry.getValue());
                 }
+            }
+
+            // Register any secure transports if available
+            if (secureTransportSettingsProviders.isEmpty() == false) {
+                final SecureTransportSettingsProvider secureSettingProvider = secureTransportSettingsProviders.iterator().next();
 
                 final Map<String, Supplier<Transport>> secureTransportFactory = plugin.getSecureTransports(
                     settings,